+++ Assets/Scripts/Editor/UnityReloaderWelcomeScreen.cs
@@ -27,7 +27,7 @@ namespace UnityReloader.Editor
             //WARN: the URL can sometimes be adjusted, make sure updated correctly
             return $"{BaseUrl}/updates/fast-script-reload/{userId}?CurrentVersion={versionId}";
         }
-        public static string VersionId = "1.0.1";
+        public static string VersionId = "1.0.2";
         private static readonly string ProjectIconName = "ProductIcon64";
         public static readonly string ProjectName = "fast-script-reload";